+++ packages/pv-scripts/webpack/base/settings/baseSettings.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ignoreWarnings: [/Failed to parse source map/],
+};