+++ packages/pv-stylemark/CHANGELOG.md
@@ -3,6 +3,91 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.8](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-stylemark@3.0.7...@pro-vision/pv-stylemark@3.0.8) (2022-01-19)
+
+**Note:** Version bump only for package @pro-vision/pv-stylemark
+
+
+
+
+
+## [3.0.7](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-stylemark@3.0.6...@pro-vision/pv-stylemark@3.0.7) (2021-11-15)
+
+
+### Bug Fixes
+
+* **pv-stylemark:** use glob pattern to target files instead of directory to work around windows bug ([b14287e](https://github.com/pro-vision/fe-tools/commit/b14287e4e3876a8a013b3d9553c05a0a6b0bc468))
+
+
+
+
+
+## [3.0.6](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-stylemark@3.0.5...@pro-vision/pv-stylemark@3.0.6) (2021-09-14)
+
+**Note:** Version bump only for package @pro-vision/pv-stylemark
+
+
+
+
+
+## [3.0.5](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-stylemark@3.0.4...@pro-vision/pv-stylemark@3.0.5) (2021-09-14)
+
+
+### Bug Fixes
+
+* **pv-stylemark:** fixed wrong paths in dev and prod scripts ([27da1d9](https://github.com/pro-vision/fe-tools/commit/27da1d9ea44047b463babfc2d2938147102479a4))
+
+
+
+
+
+## [3.0.4](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-stylemark@3.0.3...@pro-vision/pv-stylemark@3.0.4) (2021-02-24)
+
+
+### Bug Fixes
+
+* **pv-stylemark:** removed flat mathod and Assemble Instance ([85b6ca8](https://github.com/pro-vision/fe-tools/commit/85b6ca8cd80dc92081719f87620ac2ae48e01ab1))
+
+
+
+
+
+## [3.0.3](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-stylemark@3.0.2...@pro-vision/pv-stylemark@3.0.3) (2021-02-24)
+
+**Note:** Version bump only for package @pro-vision/pv-stylemark
+
+
+
+
+
+## [3.0.2](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-stylemark@3.0.1...@pro-vision/pv-stylemark@3.0.2) (2021-02-24)
+
+**Note:** Version bump only for package @pro-vision/pv-stylemark
+
+
+
+
+
+## [3.0.1](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-stylemark@3.0.0...@pro-vision/pv-stylemark@3.0.1) (2021-02-24)
+
+
+### Bug Fixes
+
+* **pv-stylemark:** add missing watchers for handlebars helpers and other assets for lsg ([5fb4d9a](https://github.com/pro-vision/fe-tools/commit/5fb4d9a0a10d81b368048ea5142dad7c9e88ebc6))
+* **pv-stylemark:** update for webpack v5 API ([3e56997](https://github.com/pro-vision/fe-tools/commit/3e56997fa74c90b75e0b6d5b3f094c4fbbfae960))
+
+
+### Performance Improvements
+
+* **pv-stylemark:** move adding files to watch list to the emit phase ([43b2485](https://github.com/pro-vision/fe-tools/commit/43b2485484c203682e37a01542d4cc76ea9fdf42))
+* **pv-stylemark:** only run stylemark jobs when neccesery ([005b4dc](https://github.com/pro-vision/fe-tools/commit/005b4dc73a742dd26775b3426f0ae2adcb7642e5))
+* **pv-stylemark:** use async spawn to run assemble scripts in parallel ([016ffd5](https://github.com/pro-vision/fe-tools/commit/016ffd5ec8c934a977eaa173d5a2ae4c058f4fd1))
+* **pv-stylemark/assemble-lite:** add new variation of assemble-lite which has a statefull instance ([ebf78b6](https://github.com/pro-vision/fe-tools/commit/ebf78b6216d46a36ec615d1f9f26f5a959fe039e))
+
+
+
+
+
 # [3.0.0](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-stylemark@2.0.6...@pro-vision/pv-stylemark@3.0.0) (2020-10-20)